video: Make all frame_number guint32
Unifies the code and ensures that:
* subclasses needing to use the frame_number on a void* field will
  always work
* wraparounds will be automatically taken care of if we have to deal
  with more than 2**32 frames

+++ gst-libs/gst/video/gstvideodecoder.c
@@ -372,8 +372,8 @@ struct _GstVideoDecoderPrivate
   int reorder_depth;
   int distance_from_sync;
 
-  guint64 system_frame_number;
-  guint64 decode_frame_number;
+  guint32 system_frame_number;
+  guint32 decode_frame_number;
 
   GList *frames;                /* Protected with OBJECT_LOCK */
   GstVideoCodecState *input_state;
